refactor(xgo): improve AST and TypeInfo architecture
- Add `xgo/ast.go` and `xgo/typeinfo.go` to enhance system abstraction.
- Replace `typesutil.Info` with improved `xgo.TypeInfo` across all
components.
- Reorganize xgo package structure with `cache.go` and `pkgdoc.go`
modules.
- Rename `proj.go` to `project.go` and update related test files.
- Update import statements from `goplus/xgo/x/typesutil` to internal
`xgo`.
- Refactor analysis passes and LSP server to use unified type system.
- Consolidate type information and AST handling for better
maintainability.
