+++ xgo/xgoutil/file.go
@@ -25,31 +25,52 @@ import (
 
 // PosFilename returns the filename for the given position.
 func PosFilename(proj *xgo.Project, pos token.Pos) string {
+	if proj == nil || !pos.IsValid() {
+		return ""
+	}
 	return proj.Fset.Position(pos).Filename
 }
 
 // NodeFilename returns the filename for the given node.
 func NodeFilename(proj *xgo.Project, node ast.Node) string {
+	if proj == nil || node == nil {
+		return ""
+	}
 	return PosFilename(proj, node.Pos())
 }
 
 // PosTokenFile returns the token file for the given position.
 func PosTokenFile(proj *xgo.Project, pos token.Pos) *token.File {
+	if proj == nil || !pos.IsValid() {
+		return nil
+	}
 	return proj.Fset.File(pos)
 }
 
 // NodeTokenFile returns the token file for the given node.
 func NodeTokenFile(proj *xgo.Project, node ast.Node) *token.File {
+	if proj == nil || node == nil {
+		return nil
+	}
 	return PosTokenFile(proj, node.Pos())
 }
 
 // PosASTFile returns the AST file for the given position.
 func PosASTFile(proj *xgo.Project, pos token.Pos) *ast.File {
+	if proj == nil || !pos.IsValid() {
+		return nil
+	}
 	astPkg, _ := proj.ASTPackage()
+	if astPkg == nil {
+		return nil
+	}
 	return astPkg.Files[PosFilename(proj, pos)]
 }
 
 // NodeASTFile returns the AST file for the given node.
 func NodeASTFile(proj *xgo.Project, node ast.Node) *ast.File {
+	if proj == nil || node == nil {
+		return nil
+	}
 	return PosASTFile(proj, node.Pos())
 }
